requirements.txt
numpy
matplotlib
scikit-learn
Pillow
scikit-image
scipy